+++ src/main/java/chess/dao/db/symbol/FileSymbol.java
@@ -1,4 +1,4 @@
-package chess.service.dao.db.symbol;
+package chess.dao.db.symbol;
 
 import chess.domain.position.File;
 import java.util.Arrays;
