+++ src/main/java/chess/dto/ChessDTO.java
@@ -1,4 +1,4 @@
-package chess.service.dto;
+package chess.dto;
 
 import chess.domain.board.Board;
 import chess.domain.piece.Color;
