+++ src/test/java/chess/service/ChessServiceTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import chess.dao.ChessDAO;
 import chess.domain.board.Board;
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
@@ -13,8 +14,7 @@ import chess.domain.piece.PieceType;
 import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
-import chess.service.dao.ChessDAO;
-import chess.service.dto.ChessDTO;
+import chess.dto.ChessDTO;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
